resource movement
