+++ pkg/appStore/deployment/service/InstalledAppService.go
@@ -48,11 +48,7 @@ import (
 	util2 "github.com/devtron-labs/devtron/pkg/util"
 	util3 "github.com/devtron-labs/devtron/util"
 	"github.com/devtron-labs/devtron/util/argo"
-	"github.com/tidwall/gjson"
 	"net/http"
-	"sync"
-	"sync/atomic"
-
 	/* #nosec */
 	"crypto/sha1"
 	"encoding/json"
@@ -91,11 +87,14 @@ type InstalledAppService interface {
 	DeployDefaultChartOnCluster(bean *cluster2.ClusterBean, userId int32) (bool, error)
 	FindAppDetailsForAppstoreApplication(installedAppId, envId int) (bean2.AppDetailContainer, error)
 	UpdateInstalledAppVersionStatus(application *v1alpha1.Application) (bool, error)
-	FetchResourceTree(rctx context.Context, cn http.CloseNotifier, appDetailsContainer *bean2.AppDetailsContainer, installedApp repository2.InstalledApps, helmReleaseInstallStatus string, status string) error
 	MarkGitOpsInstalledAppsDeletedIfArgoAppIsDeleted(installedAppId int, envId int) error
 	CheckAppExistsByInstalledAppId(installedAppId int) (*repository2.InstalledApps, error)
-	FetchChartNotes(installedAppId int, envId int, token string, checkNotesAuth func(token string, appName string, envId int) bool) (string, error)
+
 	FetchResourceTreeWithHibernateForACD(rctx context.Context, cn http.CloseNotifier, appDetail *bean2.AppDetailContainer) bean2.AppDetailContainer
+	FetchResourceTree(rctx context.Context, cn http.CloseNotifier, appDetailsContainer *bean2.AppDetailsContainer, installedApp repository2.InstalledApps, helmReleaseInstallStatus string, status string) error
+
+	//move to notes service
+	FetchChartNotes(installedAppId int, envId int, token string, checkNotesAuth func(token string, appName string, envId int) bool) (string, error)
 }
 
 type InstalledAppServiceImpl struct {
@@ -1073,155 +1072,4 @@ func (impl InstalledAppServiceImpl) CheckAppExistsByInstalledAppId(installedAppI
 		return nil, err
 	}
 	return installedApp, err
-}
-
-func (impl InstalledAppServiceImpl) FetchResourceTreeWithHibernateForACD(rctx context.Context, cn http.CloseNotifier, appDetail *bean2.AppDetailContainer) bean2.AppDetailContainer {
-	ctx, cancel := context.WithCancel(rctx)
-	if cn != nil {
-		go func(done <-chan struct{}, closed <-chan bool) {
-			select {
-			case <-done:
-			case <-closed:
-				cancel()
-			}
-		}(ctx.Done(), cn.CloseNotify())
-	}
-	acdToken, err := impl.argoUserService.GetLatestDevtronArgoCdUserToken()
-	if err != nil {
-		impl.logger.Errorw("error in getting acd token", "err", err)
-		return *appDetail
-	}
-	ctx = context.WithValue(ctx, "token", acdToken)
-	defer cancel()
-	deploymentAppName := fmt.Sprintf("%s-%s", appDetail.AppName, appDetail.EnvironmentName)
-	resourceTree, err := impl.fetchResourceTreeForACD(rctx, cn, appDetail.InstalledAppId, appDetail.EnvironmentId, appDetail.ClusterId, deploymentAppName, appDetail.Namespace)
-	appDetail.ResourceTree = resourceTree
-	if err != nil {
-		return *appDetail
-	}
-	if appDetail.ResourceTree["nodes"] == nil {
-		return *appDetail
-	}
-	appDetail.ResourceTree, _ = impl.checkHibernate(appDetail.ResourceTree, deploymentAppName, ctx)
-	return *appDetail
-}
-func (impl InstalledAppServiceImpl) checkHibernate(resp map[string]interface{}, deploymentAppName string, ctx context.Context) (map[string]interface{}, string) {
-
-	if resp == nil {
-		return resp, ""
-	}
-	responseTree := resp
-	var canBeHibernated uint64 = 0
-	var hibernated uint64 = 0
-	responseTreeNodes, ok := responseTree["nodes"]
-	if !ok {
-		return resp, ""
-	}
-	replicaNodes := impl.filterOutReplicaNodes(responseTreeNodes)
-	batchSize := impl.aCDAuthConfig.ResourceListForReplicasBatchSize
-	requestsLength := len(replicaNodes)
-	for i := 0; i < requestsLength; {
-		//requests left to process
-		remainingBatch := requestsLength - i
-		if remainingBatch < batchSize {
-			batchSize = remainingBatch
-		}
-		var wg sync.WaitGroup
-		for j := 0; j < batchSize; j++ {
-			wg.Add(1)
-			go func(j int) {
-				defer wg.Done()
-				canBeHibernatedFlag, hibernatedFlag := impl.processReplicaNodeForHibernation(replicaNodes[i+j], deploymentAppName, ctx)
-				if canBeHibernatedFlag {
-					atomic.AddUint64(&canBeHibernated, 1)
-				}
-				if hibernatedFlag {
-					atomic.AddUint64(&hibernated, 1)
-				}
-			}(j)
-		}
-		wg.Wait()
-		i += batchSize
-	}
-
-	status := ""
-	if hibernated > 0 && canBeHibernated > 0 {
-		if hibernated == canBeHibernated {
-			status = appStatus.HealthStatusHibernating
-		} else if hibernated < canBeHibernated {
-			status = appStatus.HealthStatusPartiallyHibernated
-		}
-	}
-
-	return responseTree, status
-}
-
-func (impl InstalledAppServiceImpl) processReplicaNodeForHibernation(node interface{}, deploymentAppName string, ctx context.Context) (bool, bool) {
-	currNode := node.(interface{}).(map[string]interface{})
-	resName := util3.InterfaceToString(currNode["name"])
-	resKind := util3.InterfaceToString(currNode["kind"])
-	resGroup := util3.InterfaceToString(currNode["group"])
-	resVersion := util3.InterfaceToString(currNode["version"])
-	resNamespace := util3.InterfaceToString(currNode["namespace"])
-	rQuery := &application.ApplicationResourceRequest{
-		Name:         &deploymentAppName,
-		ResourceName: &resName,
-		Kind:         &resKind,
-		Group:        &resGroup,
-		Version:      &resVersion,
-		Namespace:    &resNamespace,
-	}
-	canBeHibernatedFlag := false
-	alreadyHibernated := false
-
-	if currNode["parentRefs"] == nil {
-		canBeHibernatedFlag, alreadyHibernated = impl.checkForHibernation(ctx, rQuery, currNode)
-	}
-	return canBeHibernatedFlag, alreadyHibernated
-}
-
-func (impl InstalledAppServiceImpl) checkForHibernation(ctx context.Context, rQuery *application.ApplicationResourceRequest, currNode map[string]interface{}) (bool, bool) {
-	t0 := time.Now()
-	canBeHibernated := false
-	alreadyHibernated := false
-	ctx, _ = context.WithTimeout(ctx, 60*time.Second)
-	res, err := impl.acdClient.GetResource(ctx, rQuery)
-	if err != nil {
-		impl.logger.Errorw("error getting response from acdClient", "request", rQuery, "data", res, "timeTaken", time.Since(t0), "err", err)
-		return canBeHibernated, alreadyHibernated
-	}
-	if res.Manifest != nil {
-		manifest, _ := gjson.Parse(*res.Manifest).Value().(map[string]interface{})
-		replicas := util3.InterfaceToMapAdapter(manifest["spec"])["replicas"]
-		if replicas != nil {
-			currNode["canBeHibernated"] = true
-			canBeHibernated = true
-		}
-		annotations := util3.InterfaceToMapAdapter(manifest["metadata"])["annotations"]
-		if annotations != nil {
-			val := util3.InterfaceToMapAdapter(annotations)["hibernator.devtron.ai/replicas"]
-			if val != nil {
-				if util3.InterfaceToString(val) != "0" && util3.InterfaceToFloat(replicas) == 0 {
-					currNode["isHibernated"] = true
-					alreadyHibernated = true
-				}
-			}
-		}
-	}
-	return canBeHibernated, alreadyHibernated
-}
-
-func (impl InstalledAppServiceImpl) filterOutReplicaNodes(responseTreeNodes interface{}) []interface{} {
-	resourceListForReplicas := impl.aCDAuthConfig.ResourceListForReplicas
-	entries := strings.Split(resourceListForReplicas, ",")
-	resourceListMap := util3.ConvertStringSliceToMap(entries)
-	var replicaNodes []interface{}
-	for _, node := range responseTreeNodes.(interface{}).([]interface{}) {
-		currNode := node.(interface{}).(map[string]interface{})
-		resKind := util3.InterfaceToString(currNode["kind"])
-		if _, ok := resourceListMap[resKind]; ok {
-			replicaNodes = append(replicaNodes, node)
-		}
-	}
-	return replicaNodes
 }
